chore: lint settings and move main folders to src
