+++ src/pages/index.js
@@ -1,9 +1,9 @@
 import { useSelector } from 'react-redux';
-import { getUserData } from 'axios/User';
+import { getUserData } from 'src/axios/User';
 
 export default function Home() {
   let state = useSelector(state => state);
-  getUserData().then(res => console.log(res));
+  // getUserData().then(res => console.log(res));
   return (
     <div>
       <main>
