+++ src/pages/page1/index.js
